refact MemoCreateLayout
