+++ src/components/pages/MemoAdd.tsx
@@ -1,7 +1,35 @@
-import type { FC } from "react";
+import { FC, useState } from "react";
+import { MemoCreateLayout } from "../templates/MemoCreateLayout";
 
-type Props = {};
+type FileInfo = {
+  id: number;
+  name: string;
+};
 
-export const MemoAdd: FC<Props> = (props) => {
-  return <div>/memo/add</div>;
+type MemoContent = {
+  id: number;
+  colorCode: string;
+  tagName: string;
+  comment: string;
+  url: string;
+  createdAt: string;
+  fileInfo: FileInfo;
+};
+
+const defaultValue = {
+  id: 1,
+  colorCode: "00A8A6",
+  tagName: "",
+  comment: "",
+  url: "",
+  createdAt: "",
+  fileInfo: {
+    id: 1,
+    name: "",
+  },
+};
+
+export const MemoAdd: FC = () => {
+  const [memoContent, setMemoContent] = useState<MemoContent>(defaultValue);
+  return <MemoCreateLayout {...memoContent} />;
 };
